server/app/db/db_setup.py
import os
from sqlalchemy import create_engine
from sqlalchemy.ext.declarative import declarative_base
from sqlalchemy.orm import sessionmaker
from dotenv import load_dotenv


load_dotenv()

DATABASE_URL = os.getenv("DATABASE_URL").replace('postgres://', 'postgresql://')

engine = create_engine(
    DATABASE_URL, connect_args={}, future=True
)
SessionLocal = sessionmaker(autocommit=False, bind=engine, future=True)

Base = declarative_base()

def get_db():
    db = SessionLocal()
    try:
        yield db
    finally:
        db.close()