server/requirements.txt
alembic==1.8.1
anyio==3.6.1
asyncpg==0.26.0
bcrypt==4.0.0
click==8.1.3
databases==0.6.1
fastapi==0.85.0
fastapi-pagination==0.10.0
FastAPI-SQLAlchemy==0.2.1
gunicorn==20.1.0
h11==0.14.0
httptools==0.5.0
idna==3.4
Mako==1.2.3
MarkupSafe==2.1.1
passlib==1.7.4
psycopg2==2.9.3
pydantic==1.10.2
python-dotenv==0.21.0
PyYAML==6.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.20.4
typing_extensions==4.3.0
uvicorn==0.18.3
uvloop==0.17.0
watchfiles==0.17.0
websockets==10.3
